WCSFit: add rejection off-ramp
The fit runs until nothing gets rejected, but this causes the fit
to take (almost) forever when there are a lot of sources, because
there's always something to reject.

This commit introduces a 'clipFraction' parameter, which is the
minimum fraction of rejected sources that triggers a new iteration.
I am setting the default to zero in order to reproduce the current
behaviour, but I recommend using a value of 1.0e-2, which worked
for me when processing HSC open-use data and hitting this problem.